Moved all components into  one file
